feat: good progress
